export data to file added
