+++ lib/index.js
@@ -14,6 +14,7 @@ import { escapeHtml } from './utils/escape-html';
 import libraries from './libraries';
 import { extendDeep } from './utils/extend-deep';
 import exportImage from './utils/export-svg';
+import exportData from './utils/export-data';
 import renderDownloadButton from './utils/render-download-btn';
 import renderExecutionMetadata from './utils/render-execution-metadata';
 import copyToClipboard from './utils/copy-to-clipboard';
@@ -885,6 +886,13 @@ Dataviz.prototype.exportImage = function({ quality = 0, bgcolor = '#fff' } = {})
     node: this.config.containerElement,
     quality,
     bgcolor,
+  });
+};
+
+Dataviz.prototype.exportData = function(type='json') {
+  exportData({
+    data: this.dataset.matrix,
+    type,
   });
 };
