+++ lib/utils/download-results.js
@@ -1,22 +1,5 @@
 import exportImage from './export-svg';
-
-function generateCsvContent(data) {
-  let csvContent = 'data:text/csv;charset=utf-8,';
-
-  data.forEach((row, i) => {
-    row.forEach((cell, j) => {
-      csvContent += String(cell).replace(/,/g, '');
-      if (row.length > j + 1) {
-        csvContent += ',';
-      }
-    });
-    if (data.length > i + 1) {
-      csvContent += '\n';
-    }
-  });
-
-  return csvContent;
-}
+import exportData from './export-data';
 
 export default function downloadResults(obj) {
   const { event, type, data } = obj;
@@ -48,18 +31,10 @@ export default function downloadResults(obj) {
   }
 
   if (format === 'json') {
-    content = `data:text/json;charset=utf-8, ${encodeURIComponent(JSON.stringify(data))}`;
-    fileFormat = format;
+    exportData({ type: format, data });
   }
 
   if (format === 'csv') {
-    content = generateCsvContent(data);
+    exportData({ type: format, data });
   }
-
-  const htmlElement = document.createElement('a');
-  htmlElement.setAttribute('href', content);
-  htmlElement.setAttribute('download', `${fileName}.${fileFormat}`);
-  document.body.appendChild(htmlElement);
-  htmlElement.click();
-  document.body.removeChild(htmlElement);
 }
