+++ lib/utils/render-download-btn.js
@@ -4,7 +4,7 @@ export default function renderDownloadButton({
   element,
   label = 'Download',
   type = 'json',
-  data
+  data,
 } = {}) {
   const btnClass = 'keen-dataviz-button';
   const isBtnRendered = [...element.parentNode.children]
